# openapi.yaml
openapi: 3.0.0
info:
  title: GroceryBot API
  description: "Interact with GroceryBot directly through its API at `https://api.grocerybot.net`."
  termsOfService: https://grocerybot.net/privacy-policy/
  # contact:
  #   email: hello@swagger.io
  # license:
  #   name: Apache 2.0
  #   url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: v2.12.x
externalDocs:
  description: GroceryBot's Website
  url: https://grocerybot.net
servers:
  - url: https://api.grocerybot.net
security:
  - basicApiClient: []
  - bearerAuth: []
tags:
  - name: Auth
    description: |
      **Unstable — internal use.** These endpoints are not covered by public API stability guarantees; request and response shapes may change without a major version bump.

      Discord OAuth token exchange, JWT refresh, and session logout.
paths:
  /auth/token:
    post:
      tags: [Auth]
      summary: Exchange Discord OAuth code for GroceryBot tokens
      description: "**Unstable (internal):** Not covered by public API stability guarantees; breaking changes may occur without a major version bump.\n\nAccepts the authorization `code`, PKCE `code_verifier`, and `redirect_uri` from the Discord OAuth flow (client-driven). Validates `redirect_uri` against the server allowlist, exchanges with Discord, and returns a GroceryBot JWT plus refresh token."
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenExchangeRequest"
      responses:
        "200":
          description: Session issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Missing fields, invalid JSON, or `redirect_uri` is not allowed.
        "500":
          description: JWT issuance, encryption, or persistence failed.
        "502":
          description: Discord token exchange, `@me`, or upstream failed.
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh GroceryBot access token
      description: "**Unstable (internal):** Not covered by public API stability guarantees; breaking changes may occur without a major version bump.\n\nExchanges a valid refresh token for a new access JWT and a rotated refresh token. Returns `expires_in` for the access token (seconds)."
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: New tokens issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "400":
          description: Invalid request body.
        "401":
          description: Missing, invalid, or expired refresh token.
        "500":
          description: JWT issuance or rotation failed.
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out (revoke session)
      description: "**Unstable (internal):** Not covered by public API stability guarantees; breaking changes may occur without a major version bump.\n\nDeletes the authenticated user's session(s). Requires `Authorization: Bearer` with the GroceryBot JWT. Does not use `X-Guild-ID`."
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session removed.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          description: Server failed to delete sessions.
  /guilds:
    get:
      summary: List guilds shared by the user and bot
      description: "Returns Discord guilds for the authenticated user that the bot is also in. Requires `Authorization: Bearer`. Does not use `X-Guild-ID`."
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Guild list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGuildsResponse"
        "401":
          description: Invalid JWT, or Discord session expired — re-authenticate.
        "500":
          description: Bot session unavailable for guild intersection.
        "502":
          description: Discord `@me/guilds` failed.
  /grocery-lists:
    get:
      # tags:
      #   - grocery
      summary: GET Grocery Lists and Entries
      description: "Get your server's grocery lists and entries. Note: you have to map the relationship between a grocery list and the groceries themselves."
      parameters:
        - $ref: "#/components/parameters/XGuildIDHeader"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildGroceryList"
        "400":
          description: Bearer requests require `X-Guild-ID`.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
    post:
      summary: POST Grocery List
      description: "Create a new grocery list for your server."
      parameters:
        - $ref: "#/components/parameters/XGuildIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroceryListRequest"
      responses:
        "201":
          description: A new grocery list has been created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryList"
        "400":
          description: Bearer requests require `X-Guild-ID`; validation failure, duplicate label, or list limit reached.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /grocery-lists/{id}:
    delete:
      summary: DELETE Grocery List
      description: "Delete a grocery list by its ID. The list must have no grocery entries."
      parameters:
        - $ref: "#/components/parameters/XGuildIDHeader"
        - name: id
          in: path
          required: true
          description: The ID of the grocery list to delete.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: The grocery list has been successfully deleted.
        "400":
          description: Bearer requests require `X-Guild-ID`; or invalid ID format.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: Grocery list not found in this guild.
        "409":
          description: Grocery list still has grocery entries.
    patch:
      summary: PATCH Grocery List
      description: "Update a grocery list's display name (`fancy_name`)."
      parameters:
        - $ref: "#/components/parameters/XGuildIDHeader"
        - name: id
          in: path
          required: true
          description: The ID of the grocery list to update.
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroceryListRequest"
      responses:
        "200":
          description: The grocery list has been successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroceryList"
        "400":
          description: Bearer requests require `X-Guild-ID`; or invalid ID format or request body.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: Grocery list not found in this guild.
  /groceries:
    delete:
      summary: BATCH DELETE Grocery Entries
      description: "Delete multiple grocery entries by primary key IDs in one request (at most 300 IDs). IDs must belong to the guild selected by credentials (`X-Guild-ID` for Bearer). Duplicate IDs in the request are ignored. If any ID does not exist in this guild, the request fails with 404 and nothing is deleted."
      parameters:
        - $ref: "#/components/parameters/XGuildIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroceryBatchDeleteRequest"
      responses:
        "204":
          description: All requested grocery entries were deleted.
        "400":
          description: Invalid JSON, missing `ids`, empty `ids`, more than 300 IDs, or invalid ID values.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: One or more grocery entry IDs were not found in this guild; no entries were deleted.
    post:
      # tags:
      #   - grocery
      summary: POST Grocery Entry
      description: "Create a new grocery entry for your server. To attach it to a grocery list, obtain your server's grocery lists from GET /grocery-lists and put it into `grocery_list_id`."
      parameters:
        - $ref: "#/components/parameters/XGuildIDHeader"
      requestBody:
        description: Creates a new grocery entry.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroceryEntry"
      responses:
        "201":
          description: A new grocery entry has been created.
        "400":
          description: Validation fails. You must ensure that `id` is empty.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /groceries/{id}:
    delete:
      # tags:
      #   - grocery
      summary: DELETE Grocery Entry
      description: "Delete a grocery entry by its ID. You can obtain the ID from the GET /grocery-lists endpoint."
      parameters:
        - $ref: "#/components/parameters/XGuildIDHeader"
        - name: id
          in: path
          required: true
          description: The ID of the grocery entry to delete.
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: The grocery entry has been successfully deleted.
        "400":
          description: Invalid ID format.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          description: Grocery entry not found.
  /registrations:
    get:
      summary: GET guild registrations
      description: Returns active registrations for the guild selected by credentials (Basic scope) or by `X-Guild-ID` (Bearer).
      parameters:
        - $ref: "#/components/parameters/XGuildIDHeader"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuildRegistration"
        "400":
          description: Bearer requests require `X-Guild-ID`.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          description: Server error.
components:
  parameters:
    XGuildIDHeader:
      name: X-Guild-ID
      in: header
      required: true
      description: "Required for Bearer-authenticated requests to guild-scoped routes (see SPEC-001). Ignored for Basic auth (guild is determined from the API client scope)."
      schema:
        type: string
  schemas:
    TokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri]
      properties:
        code:
          type: string
          description: Authorization code from Discord OAuth redirect.
        code_verifier:
          type: string
          description: PKCE code verifier (S256) used in the authorization request.
        redirect_uri:
          type: string
          description: Must match the redirect URI used with Discord and must be allowed on the server.
    TokenResponse:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
          description: "GroceryBot JWT (use as Authorization Bearer header)."
        refresh_token:
          type: string
          description: "Opaque refresh token for POST /auth/refresh."
        expires_in:
          type: integer
          format: int64
          description: Access token lifetime in seconds.
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    RefreshTokenResponse:
      $ref: "#/components/schemas/TokenResponse"
    UserGuild:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
          nullable: true
    UserGuildsResponse:
      type: object
      properties:
        guilds:
          type: array
          items:
            $ref: "#/components/schemas/UserGuild"
    GuildRegistration:
      type: object
      description: "Wire format uses exported Go field names (default `encoding/json`). Nested entitlements may be present when preloaded."
      properties:
        ID:
          type: integer
        GuildID:
          type: string
        RegistrationEntitlementID:
          type: integer
        RegistrationEntitlement:
          type: object
          nullable: true
        CreatedAt:
          type: string
          format: date-time
        ExpiresAt:
          type: string
          format: date-time
          nullable: true
    GuildGroceryList:
      type: object
      description: Represents a server's grocery list information, including groceries and entries. They are unordered, so you must map the groceries <-> grocery lists relationship yourself.
      properties:
        guild_id:
          type: string
          description: The server ID that this grocery list information represents.
        grocery_entries:
          type: array
          items:
            $ref: "#/components/schemas/GroceryEntry"
        grocery_lists:
          type: array
          items:
            $ref: "#/components/schemas/GroceryList"
    GroceryBatchDeleteRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 300
          items:
            type: integer
            minimum: 1
          description: Primary keys of grocery entries to delete (see GET /grocery-lists). Values are unsigned (positive integers). At most 300 IDs per request.
    CreateGroceryListRequest:
      type: object
      required: [list_label]
      properties:
        list_label:
          type: string
          description: "The unique label used to refer to this grocery list in commands (e.g. `gro:amazon`)."
        fancy_name:
          type: string
          nullable: true
          description: Optional display name shown alongside the label.
    UpdateGroceryListRequest:
      type: object
      required: [fancy_name]
      properties:
        fancy_name:
          type: string
          nullable: true
          description: Display name for the grocery list. Send `null` to clear it.
    GroceryEntry:
      type: object
      description: Represents a grocery entry (e.g. added by /gro).
      required: [item_desc]
      properties:
        id:
          type: number
          description: Primary key of the entry. Guaranteed to be unique.
          readOnly: true
        created_at:
          type: string
          description: A timestamp on when the entry was created. Automatically generated - you cannot assign this manually.
          readOnly: true
        updated_at:
          type: string
          description: A timestamp on when the entry was updated. Automatically generated - you cannot assign this manually.
          readOnly: true
        item_desc:
          type: string
          description: "Description of the entry. For example, `/gro chicken` would result in this field being `chicken`."
        updated_by_id:
          type: string
          description: Discord ID of the user who updated this entry last.
          deprecated: true
          readOnly: true
        grocery_list_id:
          type: integer
          format: int64
          description: The foreign key pointing to a particular grocery list in your server. A null value means that this belongs to the server's default grocery list.
          nullable: true
    GroceryList:
      type: object
      description: Represents a grocery list (e.g. added by /grolist-new).
      required: [id, guild_id, list_label]
      properties:
        id:
          type: number
          description: Primary key of the list. Guaranteed to be unique.
          readOnly: true
        created_at:
          type: string
          description: A timestamp on when the list was created.
          readOnly: true
        updated_at:
          type: string
          description: A timestamp on when the list was updated.
          readOnly: true
        guild_id:
          type: string
          description: "The server ID to which the list belongs to."
        list_label:
          type: string
          description: "The unique label that is used to refer to this grocery list in commands. For example, `!gro:ebay blah` would be referring to a grocery list with `list_label` as \"blah\"."
        fancy_name:
          type: string
          description: The fancy name that is displayed to the users (alongside `list_label`) when displaying this grocery list in commands.
          nullable: true

  # requestBodies:
  #   Pet:
  #     description: Pet object that needs to be added to the store
  #     content:
  #       application/json:
  #         schema:
  #           $ref: "#/components/schemas/Pet"
  #       application/xml:
  #         schema:
  #           $ref: "#/components/schemas/Pet"
  #   UserArray:
  #     description: List of user object
  #     content:
  #       application/json:
  #         schema:
  #           type: array
  #           items:
  #             $ref: "#/components/schemas/User"
  responses:
    UnauthorizedError:
      description: API credentials are incorrect. These must be generated by running GroceryBot's /developer command in your Discord server.
    ForbiddenError:
      description: API credentials do not have enough permissions to perform that action. Most of these errors are left intentionally ambiguous in order to prevent abuse. Reach out to GroceryBot's Discord server if you would like to resolve a constant 403.
  securitySchemes:
    basicApiClient:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: GroceryBot JWT from POST /auth/token or POST /auth/refresh. On grocery and registration routes, send `X-Guild-ID` to select the target server.